1) Move the sprite/tileset hooks to a specific source file. 2) Account for other debugging bitmasks.
