+++ modules/military.cpp
@@ -14,7 +14,7 @@
 
 #pragma intrinsic(_ReturnAddress)
 
-#define MILITARY_DEBUG_PLACEMENT 1
+#define MILITARY_DEBUG_PLACEMENT_OTHER 1
 #define MILITARY_DEBUG_PLACEMENT_NAVAL 2
 
 #define MILITARY_DEBUG DEBUG_FLAGS_NONE
