+++ sc2kfix.vcxproj
@@ -97,6 +97,7 @@
     <ClCompile Include="hooks\hook_animation.cpp" />
     <ClCompile Include="hooks\hook_sc2kdemo_miscellaneous.cpp" />
     <ClCompile Include="hooks\hook_guzzardo.cpp" />
+    <ClCompile Include="hooks\hook_spritesandtilesets.cpp" />
     <ClCompile Include="kuroko\builtins.c" />
     <ClCompile Include="kuroko\chunk.c" />
     <ClCompile Include="kuroko\compiler.c" />
