+++ sc2kfix.vcxproj.filters
@@ -335,6 +335,9 @@
     <ClCompile Include="hooks\hook_guzzardo.cpp">
       <Filter>Source Files\Hooks</Filter>
     </ClCompile>
+    <ClCompile Include="hooks\hook_spritesandtilesets.cpp">
+      <Filter>Source Files\Hooks</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="sc2kfix.rc">
